LayoutTests/imported/w3c/web-platform-tests/fullscreen/api/document-exit-fullscreen-nested-in-iframe-expected.txt



FAIL Exit fullscreen for nested fullscreen inside an iframe assert_equals: expected Element node <div id="outer"><div id="inner"></div></div> but got null
